Remove TriggerOpen.cs
Unnecessary since Door now uses RoomClearLogic.cs
