+++ IoTether/Assets/Scripts/StageScripts/RoomClearLogic.cs
@@ -1,0 +1,42 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class RoomClearLogic : MonoBehaviour
+{
+    private GameObject Closed;
+    private GameObject Opened;
+    public Transform enemiesInRoom;
+    // Start is called before the first frame update
+    void Start()
+    {
+        Closed = gameObject.transform.GetChild(0).gameObject;
+        Opened = gameObject.transform.GetChild(1).gameObject;
+        Close();
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (enemiesInRoom.childCount == 0)
+        {
+            Open();
+        }   
+        
+    }
+
+    [ContextMenu("Close")]
+    private void Close()
+    {
+        Closed.SetActive(true);
+        Opened.SetActive(false);
+    }
+    [ContextMenu("Open")]
+    private void Open()
+    {
+        Closed.SetActive(false);
+        Opened.SetActive(true);
+        Debug.Log("Door Opened");
+    }
+}
