+++ IoTether/Assets/Scripts/StageScripts/RoomClearLogic.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 798a3adef980d854b856a617b810394b
+guid: 40c3553c7de0dc449a425c3f6df6e9d6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
