+++ IoTether/Assets/Scripts/StageScripts/TextUpdateOnPosition.cs
